Explain things a bit better in buildsrc readme
Also while trying to explain things, it took me a lot of time to explain why I had to
include incorporeal.repositories.gradle.kts from java-conventions purely to get it
in the subprojects transitively. There's no point to that. Yeah "don't repeat yourself"
and all that, but typing is free and the clarity is worth it.